httpd.py refactored
